requirements.txt
Flask==2.0.2
paho-mqtt==1.5.1
pytz==2021.3
werkzeug==2.0.2
